Replace a bunch of size_t with unsigned
I've apparently gotten into the wrong mindset about size_t vs
unsigned.  It's easy to let size_t proliferate throughout your
code base once you start matching types with all of the standard
library arguments and thinking of this as the "natural" type for
anything that counts or sizes things.  It seems to have a certain
cleanliness advantage in that it won't wrap in the real world,
too.

However, it's commonly wasteful (e.g. x86_64) when the sizes are
known to be reasonably-limited, and better to stick with unsigned
in the general case.

We've kept size_t in places where it clearly is a good idea: when
it's directly storing a size_t return value from a function before
checking and converting, for our own library functions related to
raw allocation and file-mapping, and for file data offsets in
readers of potentially-large files (like the zonefile scanner).