Replace JKISS32 with JKISS
JKISS is a bit higher in quality than JKISS32 (although still not
cryptographic, which is fine!) at the cost of a couple multiplies,
and it's really not a big speed hit to use it.  Also, I'm a little
more confident in how we seed this variant properly to avoid edge
cases.  We have more use-cases coming and I'm starting to care
more about its quality affecting algorithms whose properties rely
on "random" choices.