Experimental dnspython-based testsuite infra
This is a basic ugly stab at a python3 + pytest + dnspython
testsuite alternative.  Its structure loosely mirrors the
existing Perl testsuite, but most of the advanced and/or hacky
stuff has been left out for now while figuring out the basic shape
of the solution space and how it integrates with autoconf and the
tree, how to best make use of various fixtures and scopes, etc.

Using tox is a tempting way to simplify integration and
dependencies, but I'm hesitant to do that because it would require
builders to fetch network stuff and would run up against e.g.
debian policy issues.

I'm not yet far enough along with this to decide if it's a viable
path.  There are other competing ideas, including:

* Keep python3, but swap in dnslib for dnspython, or use both
  together if they have different strengths for corner cases.
* Go + https://github.com/miekg/dns ?
* A C-based solution built on https://github.com/NLnetLabs/ldns
* Could also contemplate shifting a lot of the deep-detail testing
  down to unit-level testing via ldns, and then whatever our
  regression suite choice is for whole-daemon-via-network testing
  can be a lot simpler.
* Stick with the Perl + Net::DNS testsuite we have, but
  significantly refactor/cleanup/modernize.  It was at best ugly
  and poorly-factored when it was created, and it has accumulated
  a ton of cruft and debt since!

With any but the last option, we'd keep the existing Net::DNS
testsuite around in parallel for a while until we're confident
that it's completely redundant in value.

No matter which path we go down, the main issue is we need to add
DNSSEC support to the testsuite.  Doing it with the existing
Net::DNS -based suite without significantly revamping it first
would probably be a nightmare of hackery, so this is a good time
to explore other options!

It'd be nice if whatever we land on could be even more declarative
in style than the existing tests, too (they're pretty close for
most of the suite, though!).  With any of the possible infra
options, we could build a system of declarative test inputs and
outputs that are recorded as pure data...