Use C11 _Thread_local
This doesn't buy us anything fancy, but as long as we're on C11
now, it's more standards-based than the legacy __thread.

The "thread_local" variant would look nicer, but I'm not sure that
the requisite "threads.h" support has really caught on broadly yet
at the libc level (if it ever did, maybe we could even swap out
all our pthreads usage, but that's a whole other matter to
investigate).