Split up assertions and assumptions
Previously, we had only "gdnsd_assert(expr)", and aside from
documenting our assumptions for other humans, this macro served
two functional purposes together at the same time for the same set
of expressions throughout the code: In developer debug builds, it
did something akin to "assert()", and in production NDEBUG builds
it did something more like clang's "__builtin_assume()", in that
it assured compiler and/or tools of facts they may not easily be
able to derive for themselves, which might aid them in
optimization and/or analysis.

There are a few issues with these combined, inseparable
assume/assert statements:

1) The "assume" part of this can be dangerous if you're not very
   sure of the assumptions, as it can lead to mis-optimization and
   even undefined behavior.  It can be surprising sometimes, even
   after getting used to our style, that a macro that sounds like
   an assert can break things in NDEBUG builds.
2) While all assert/assume expressions *should* be side-effect
   free in practice, it can be difficult for the compiler to prove
   that's the case.  With a real "assert", the whole expression
   goes away in production builds, but with "assume"-like macros
   on GCC (which must be built from a conditional branch and a
   __builtin_unreachable), asserted expressions which can't be
   trivially proven to be side-effect free may still be executed
   unexpectedly in production builds and incur a perf penalty.
3) In many cases, it's clear that the expression has no potential
   optimization or analysis benefits for the assume case and
   really is just a debugging assertion to aid humans against
   introducing regressions with new development, in which case
   we're better off not even going down the assume road with all
   of its pitfalls for those cases.

As of this commit, the functionality has now been split up into
two separate macros "gdnsd_assert(expr)" and "gdnsd_assume(expr)".
The assert variant only acts as a better "assert()" and always
goes away completely in NDEBUG builds, while the assume variant
does what we were doing before: it's an assert in developer
builds, and a true assume in NDEBUG builds.

All of the existing gdnsd_assert() statements were initially
mechanically converted to gdnsd_assume(), as this preserves their
status quo and thus is the least-impactful thing to do.
Afterwards, I made a pass through looking at them all in detail,
and downgraded several of them back to the new assert-only
gdnsd_assert() statement where it seemed to make the most sense:
where they clearly had no benefits in the assume case and/or where
it was known that the compiler couldn't trivially prove their lack
of side-effects.

As one would expect, I found a number of other minor QA nits while
re-auditing all of those assertions, which required making some
unrelated minor changes, but those changes were all backported to
earlier QA commits to leave this one fairly mechanical and clean.

The gdnsd_assume() macro also has one upgrade over the previous
arrangement, which is that in clang builds it will take advantage
of the simpler __builtin_assume() instead of the conditional
branch towards __builtin_unreachable() we have to use on GCC.